// src/app/pages/map/interfaces/instructor-data.interfaces.ts
// export interface InstructorData {
//   id: number,
//   img: string,
//   name: string,
//   title: string,
//   courseImage: string,
//   listImg: string,
//   lesson: string,
//   rating: string,
//   teacherImg: string,
//   teacherName: string,
//   category: string,
//   price: string,
//   oldPrice: string,
//   descripcion: string;
//   lngLat: [number, number];
// }
export interface Profesores {
  Nombre: string;
  category: string;
  descripcion:string;
  rating:string;
  lngLat: [number, number];
}